LAP2.Crypto.Constructions.ClaimableRS (LAP2 v0.1.0)

This module implements a Claimable Ring Signature (CRS) scheme. The scheme allows a member to claim their signature after the fact by revealing a commitment secret.

This encryption scheme is based on the specification from the following paper:

Park, S. and Sealfon, A., 2019. It wasn’t me! Repudiability and claimability of ring signatures. In Advances in Cryptology–CRYPTO 2019: 39th Annual International Cryptology Conference, Santa Barbara, CA, USA, August 18–22, 2019, Proceedings, Part III 39 (pp. 159-190). Springer International Publishing.

Types

sag()

@type sag() :: %{chal: charlist(), ring: [charlist()], resp: [charlist()]}

Functions

crs_claim(ring_idx, arg1, arg2)

@spec crs_claim(
  non_neg_integer(),
  {{charlist(), charlist()}, charlist(), charlist(), charlist()},
  {sag(), charlist()}
) :: {:ok, {charlist(), charlist()} | :invalid_commitment} | {:error, atom()}

Generate a claim for a C-RS signature.

crs_gen()

@spec crs_gen() :: %{
  vk: {charlist(), charlist()},
  sk: {{charlist(), charlist()}, charlist(), charlist(), charlist()}
}

Generate a Claimable RS (C-RS) key pair.

crs_sign(ring_idx, arg, ring, msg)

@spec crs_sign(
  non_neg_integer(),
  {{charlist(), charlist()}, charlist(), charlist(), charlist()},
  [charlist()],
  charlist()
) :: {:ok, {sag(), charlist()}} | {:error, atom()}

Generate a C-RS signature.

crs_vrfy(arg, msg)

@spec crs_vrfy({sag(), charlist()}, charlist()) :: {:ok, boolean()} | {:error, atom()}

Verify a C-RS signature.

crs_vrfy_claim(arg1, arg2, arg3)

@spec crs_vrfy_claim(
  {charlist(), charlist()},
  {sag(), charlist()},
  {charlist(), charlist()}
) ::
  {:ok, boolean()} | {:error, atom()}

Verify the validity of a claim for a C-RS signature.

rs_gen()

@spec rs_gen() :: {charlist(), charlist()}

Generate a secret and public key pair tuple for a Claimable Ring Signature (CRS) scheme.

rs_sign(ring_idx, sk, ring, msg)

@spec rs_sign(non_neg_integer(), charlist(), [charlist()], charlist()) ::
  {:error, atom()} | {:ok, sag()}

Generate an SAG ring signature for a message. This function verifies the validity of the arguments before calling the Rust NIF.

rs_vrfy(sag, msg)

@spec rs_vrfy(
  sag(),
  charlist()
) :: {:error, atom()} | {:ok, boolean()}

Verify an SAG ring signature for a message.

sag_to_charlist(sag)

@spec sag_to_charlist(sag()) :: charlist()

Merge a SAG struct to a flat charlist.