LAP2.Crypto.Constructions.CryptoNifs (LAP2 v0.1.0)

Functions

commit_gen(sk, rand)

@spec commit_gen(charlist(), charlist()) :: charlist()

Generate a hash-based (SHA256) commitment of a message.

commit_vrfy(sk, rand, com)

@spec commit_vrfy(charlist(), charlist(), charlist()) :: boolean()

Verify a hash-based (SHA256) commitment.

prf_eval(sk, data)

@spec prf_eval(charlist(), charlist()) :: charlist()

Generate pseudo-random function (PRF) output. The PRF uses AES128-CMAC.

prf_gen(n)

@spec prf_gen(non_neg_integer()) :: charlist()

Generate n-bit randomness (floored to nearest byte).

rs_nif_gen()

@spec rs_nif_gen() :: {charlist(), charlist()}

Generate a secret and public key pair tuple for a Spontaneous Anonymous Group (SAG) ring signature scheme.

rs_nif_sign(idx, sk, ring, msg)

@spec rs_nif_sign(non_neg_integer(), charlist(), [charlist()], charlist()) ::
  {charlist(), [charlist()], [charlist()]}

Sign a message using a Spontaneous Anonymous Group (SAG) ring signature.

rs_nif_vrfy(chal, ring, resp, msg)

@spec rs_nif_vrfy(charlist(), [charlist()], [charlist()], charlist()) :: boolean()

Verify a Spontaneous Anonymous Group (SAG) ring signature.

standard_signature_gen()

@spec standard_signature_gen() :: {charlist(), charlist()}

Generate a secret and public key pair tuple for a ed25519_dalek signature scheme.

standard_signature_sign(sk, msg, rand)

@spec standard_signature_sign(charlist(), charlist(), charlist()) :: charlist()

Sign a message using a ed25519_dalek signature scheme.

standard_signature_vrfy(sig, vk, msg)

@spec standard_signature_vrfy(charlist(), charlist(), charlist()) :: boolean()

Verify a ed25519_dalek signature.