LAP2.Crypto.InformationDispersal.RabinIDA (LAP2 v0.1.0)

Rabin's Information Dispersal Algorithm, used to split and reconstruct data.

Functions

decode_double_byte(bytes)

@spec decode_double_byte(binary()) :: [non_neg_integer()]

encode_double_byte(bytes)

@spec encode_double_byte([non_neg_integer()]) :: [non_neg_integer()]

reconstruct(shares)

@spec reconstruct([map()]) :: {:ok, binary()} | {:error, nil}

Reconstruct the data from the given shares. Generates the reassembly matrix and performs matrix multiplication with the dat a in the shares. Finally, the data is un-padded and returned.

split(data, n, m)

@spec split(binary(), non_neg_integer(), non_neg_integer()) :: [binary()]

Split the data into the given number of shares. n is the number of shares to split the data into. m is the number of shares required to reconstruct the data (threshold).