LAP2.Crypto.InformationDispersal.SecureIDA (LAP2 v0.1.0)

Security enhanced Information Dispersal Algorithm, used to split and reconstruct data.

TODO actually make it use a proper IDA rather than splitting the data lol

Functions

disperse(data, n, m, message_id)

@spec disperse(binary(), integer(), integer(), integer()) :: [Share.t()]

Encrypt and split the data into the given number of shares. Encryption keys are split using Shamir's secret sharing, while the data is split using Rabin's IDA.

Arguments

data - The data to split (binary)

n - The number of shares to generate (non-negative integer)

m - The size of each share (non-negative integer), also the reconstruction threshold

Returns

A list of shares, each containing the share data and its index (list(Share.t()))

reconstruct(shares)

@spec reconstruct([Share]) :: {:ok, binary()}

Reconstruct the original data from the given shares.

Arguments

shares - The shares to reconstruct the data from (list(Share.t()))

Returns

{:ok, data} - The reconstructed data (binary)

{:error, reason} - The reason why the data could not be reconstructed (binary)