LAP2.Main.ConnectionSupervisor (LAP2 v0.1.0)

Module for supervising the status of various network connections. Dispatches proxy requests via the ProxyManager. Encodes/decodes and routes data above the proxy abstraction layer. Receives commands directly from the Master module, interfaces with ProxyManager and IntroductionPoint modules. Relays are abstracted away by the ProxyManager, this module handles connections bound to proxy sequences.

Functions

child_spec(init_arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

debug(name)

deliver_inbound(data, proxy_seq, registry_table)

@spec deliver_inbound(binary(), non_neg_integer(), map()) :: :ok | :error

Deserialise and deliver an inbound Query to the appropriate route

establish_proxy(registry_table)

@spec establish_proxy(map()) :: :ok | :error

Attempt to establish a new anonymous proxy.

init(config)

@spec init(map()) :: {:ok, map()}

Callback implementation for GenServer.init/1.

modify_connection(proxy_seq, atom, service_ids, conn_supervisor)

@spec modify_connection(non_neg_integer(), atom(), [binary()], atom()) :: :ok | :error

Modify an existing connection to an introduction point.

proxy_established(proxy_seq, registry_table)

@spec proxy_established(non_neg_integer(), map()) :: :ok | :error

Proxy establishment successful, broadcast to listeners

remove_proxy(proxy_seq, registry_table)

@spec remove_proxy(non_neg_integer(), map()) :: :ok | :error

Removes a proxy connection from the ProxyManager and ConnectionSupervisor.

request_introduction_point(service_ids, registry_table)

@spec request_introduction_point([String.t()], map()) :: :ok | :error

Send an introduction point request to an existing connection. A list of service identifiers are passed in the request.

request_introduction_point(listener_id, service_ids, registry_table)

send_queries(queries, registry_table)

@spec send_queries([Query.t()], map()) :: :ok | :error

Send a list of outbound queries via random connections.

send_query(query, proxy_seq, registry_table)

@spec send_query(Query.t(), non_neg_integer(), map()) :: :ok | :error

Send a query to an introduction point/host

send_responses(response_list, registry_table)

@spec send_responses([map()], map()) :: :ok | :error

Send out a list of responses

service_lookup(service_id, name)

@spec service_lookup(String.t(), atom()) ::
  {:ok, non_neg_integer()} | {:error, atom()}

start_link(config)

@spec start_link(map()) :: GenServer.on_start()

Start the ConnectionSupervisor process.

successful_key_rotation(proxy_seq, registry_table)

@spec successful_key_rotation(non_neg_integer(), map()) :: :ok | :error

Key rotation successful, broadcast to listeners

teardown_introduction_point(proxy_seq, registry_table)

@spec teardown_introduction_point(non_neg_integer(), map()) :: :ok | :error

Send an teardown request to an existing introduction point. Delete all associated state.