LAP2.Main.Helpers.ProxyHelper (LAP2 v0.1.0)

Contains helper functions for processing common proxy requests and responses

Functions

accept_proxy_request(request, map, state)

@spec accept_proxy_request(
  Request.t(),
  %{proxy_seq: non_neg_integer(), clove_seq: non_neg_integer(), relays: list()},
  map()
) :: map()

Accept a proxy request. Respond with an acknowledgement. Add the relays to the proxy pool and return updated Proxy GenServer state.

handle_discovery_response(request, map, state)

@spec handle_discovery_response(Request.t(), map(), map()) ::
  {:ok, map()} | {:error, atom()}

Validate the length of the anonymous path.

handle_fin_key_exhange(request, proxy_seq, state)

@spec handle_fin_key_exhange(Request.t(), non_neg_integer(), map()) :: :ok | :error

Finalise the key exchange.

handle_key_exchange_ack(proxy_seq, registry_table)

@spec handle_key_exchange_ack(non_neg_integer(), map()) :: :ok | :error

handle_key_rotation(request, proxy_seq, state)

@spec handle_key_rotation(Request.t(), non_neg_integer(), map()) :: :ok | :error

Handle a key rotation request.

handle_key_rotation_ack(proxy_seq, registry_table)

@spec handle_key_rotation_ack(non_neg_integer(), map()) :: :ok | :error

Handle a key rotation acknowledgement.

handle_proxy_query(request, pseq, registry_table)

@spec handle_proxy_query(Request.t(), non_neg_integer(), map()) :: :ok | :error

Route proxy query to the correct application listener via the Master module.

init_proxy_request(registry_table, clove_seq, cloves)

@spec init_proxy_request(map(), non_neg_integer(), non_neg_integer()) :: :ok | :error

Initialise proxy request. (outbound) Generate a new EncryptedRequest struct and send it to random neighbors. @param registry_table - RegistryTable struct @param clove_seq - Clove sequence number @param cloves - Number of cloves to be sent

remove_proxy(state, proxy_seq)

@spec remove_proxy(map(), non_neg_integer()) :: map()

Remove a proxy from the proxy pool.

send_to_proxy(request, proxy_seq, relay_pool, registry_table)

@spec send_to_proxy(Request.t(), non_neg_integer(), list(), map()) :: :ok | :error

Send an outbound proxy query to a single specific proxy.