LAP2.Main.Master (LAP2 v0.1.0)

Module for sending/receiving data via the network.

Functions

bootstrap_dht(bootstrap_addr, master_name)

@spec bootstrap_dht(
  {String.t(), non_neg_integer()},
  atom()
) :: :ok | :error

Bootstrap the DHT table with a request to a known network node.

cache_query(query, routing_info, service_id, name)

@spec cache_query(Query.t(), map(), String.t(), atom()) :: :ok | :error

Add a query to the cache so it can be retrieved for a response.

child_spec(init_arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

deregister_listener(listener_id, master_name)

@spec deregister_listener(String.t(), atom()) :: :ok

Deregister a listener with a given stream ID.

deregister_service(service_id, secret, master_name)

@spec deregister_service(String.t(), String.t(), atom()) :: :ok | :error

Attempt to deregister a service with a given stream ID and its associated secret key.

discard_query(service_id, query_id, secret, name)

@spec discard_query(String.t(), non_neg_integer(), String.t(), atom()) :: :ok | :error

Delete a query from the cache. Used if a service does not wish to respond to a query.

discover_proxy(master_name)

@spec discover_proxy(atom()) :: :ok

Initialise the proxy discovery process.

get_listener_struct(listener_id, master_name)

@spec get_listener_struct(String.t(), atom()) :: {:ok, map()} | {:error, atom()}

get_registry_table(name)

@spec get_registry_table(atom()) :: map()

get_service_target(listener_id, master_name)

@spec get_service_target(String.t(), atom()) ::
  {:ok, atom() | tuple()} | {:error, atom()}

Send a response to the appropriate listener socket.

init(config)

@spec init(map()) :: {:ok, map()}

Initialise the data handler GenServer.

pop_return(query_id, name)

@spec pop_return(non_neg_integer(), atom()) :: String.t() | :error

register_dht_request(query_id, name)

@spec register_dht_request(non_neg_integer(), atom()) :: :ok

Remember a query ID for a DHT request to verify the response

register_listener(listener, master_name)

@spec register_listener(atom() | tuple(), atom()) :: {:ok, String.t()}

Register a listener with a unique stream ID. :stdout for STDOUT {:tcp, {ip, port}} for TCP {:native, funct, service_id} for a native listener function

register_return(query_id, listener_id, name)

@spec register_return(non_neg_integer(), String.t(), atom()) :: :ok

register_service(service_id, listener_id, master_name)

@spec register_service(String.t(), String.t(), atom()) ::
  {:ok, String.t()} | {:error, atom()}

Register a service with a given stream ID. This function returs a secret key needed to modify the registered service.

request_remote(intro_points, data, service_id, listener_id, name)

@spec request_remote(
  [{String.t(), non_neg_integer()}],
  binary(),
  String.t(),
  String.t(),
  atom()
) ::
  :ok | :error

respond(service_id, query_ids, secret, data, name)

@spec respond(String.t(), [non_neg_integer()], String.t(), binary(), atom()) ::
  :ok | :error

Respond to a query with a given query ID.

service_lookup(service_id, master_name)

@spec service_lookup(String.t(), atom()) :: {:ok, String.t()} | {:error, atom()}

Lookup a service to get its associated stream ID.

setup_introduction_point(service_ids, name)

@spec setup_introduction_point([String.t()], atom()) :: :ok | :error

Get the registry table for a given master name.

start_link(config)

@spec start_link(map()) :: GenServer.on_start()

Start the Master process.

valid_dht_request?(query_id, name)

@spec valid_dht_request?(non_neg_integer(), atom()) :: boolean()

Verify if a query ID is valid for a DHT request