LAP2.Networking.Helpers.State (LAP2 v0.1.0)

Module for managing the state of the routing module.

Functions

add_own_clove(state, clove_seq)

@spec add_own_clove(map(), non_neg_integer()) :: map()

Add clove to list of own cloves.

add_proxy_relay(state, proxy_seq, relay_node)

@spec add_proxy_relay(map(), non_neg_integer(), {String.t(), integer()}) :: map()

Add entry to relay table

add_relay(state, proxy_seq, relay_1, relay_2)

@spec add_relay(
  map(),
  non_neg_integer(),
  {String.t(), non_neg_integer()},
  {String.t(), non_neg_integer()}
) :: map()

ban_clove(state, clove_seq)

@spec ban_clove(map(), non_neg_integer()) :: map()

Add a drop rule for the clove_Seq.

cache_clove(state, source, dest, clove)

@spec cache_clove(map(), {binary(), integer()}, {binary(), integer()}, Clove.t()) ::
  map()

Add a clove to the local cache.

clean_state(state)

@spec clean_state(map()) :: map()

Remove outdated entries from the state based on their timestamps.

delete_own_clove(state, clove_seq)

@spec delete_own_clove(map(), non_neg_integer()) :: map()

Remove a clove from the list of own cloves.

evict_clove(state, clove_seq)

@spec evict_clove(map(), non_neg_integer()) :: map()

Evict a clove from the cache.

get_route(state, source, clove)

@spec get_route(map(), {String.t(), non_neg_integer()}, Clove.t()) ::
  atom()
  | {atom(), {String.t(), non_neg_integer()} | atom() | binary()}
  | {:discovery_response, {String.t(), non_neg_integer()},
     {String.t(), non_neg_integer()}}

Get routing information from state

remove_neighbor(state, neighbor)

@spec remove_neighbor(map(), String.t()) :: map()

Remove a neighbor from the list of random neighbors.

update_clove_timestamp(state, clove_seq)

@spec update_clove_timestamp(map(), non_neg_integer()) :: map()

Update the cache entry timestamp.

update_relay_timestamp(state, proxy_seq)

@spec update_relay_timestamp(map(), non_neg_integer()) :: map()

Update the relay table entry timestamp.