LAP2.Networking.Router (LAP2 v0.1.0)

Stores routing information, updates DHT and stores information for resolving addresses. The actual functions for resolving the addresses, extracting keys, etc. is located in LAP2.Networking.Resolver This module also stores connection information (anonymous paths and corresponidng crypto keys) TODO have a look at a secure way of storing the keys

Functions

add_proxy_relays(proxy_seq, relays, name)

@spec add_proxy_relays(non_neg_integer(), [{String.t(), non_neg_integer()}], atom()) ::
  :ok

Accept and reply to a proxy request, update state accordingly.

append_dht(lap2_addr, ip_addr, name)

@spec append_dht(binary(), {String.t(), non_neg_integer()}, atom()) :: :ok

Update the DHT with a new entry in the state

cache_query(query, new_qid, conn_id, name)

@spec cache_query(Query.t(), non_neg_integer(), String.t(), atom()) :: :ok

Cache an inbound TCP query so it can be used in the response. This is a used by Introduction Points.

child_spec(init_arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

clean_state(name)

@spec clean_state(atom()) :: :ok

Delete outdated entries from state.

debug(name)

@spec debug(atom()) :: :ok

Inspect the router state.

drop_test(route, drop_probab)

@spec drop_test(atom() | tuple(), float()) :: atom() | tuple()

get_cached_query(qid, name)

@spec get_cached_query(non_neg_integer(), atom()) ::
  {:ok, Query.t()} | {:error, atom()}

Attempt to fetch a cached query from the router's cache. If retrieved, it is deleted from the cache to conserve memory.

get_dht(name)

@spec get_dht(atom()) :: map()

Get the DHT from the state

get_neighbors(name)

@spec get_neighbors(atom()) :: [String.t()]

Get a list of neighbors from the state

get_query_relay(query_id, name)

@spec get_query_relay(non_neg_integer(), atom()) ::
  {:ok, non_neg_integer()} | {:error, atom()}

Retrieve the origin proxy sequence number for a given query ID.

init(config)

@spec init(map()) :: {:ok, map()}

Initialise the router.

route_inbound(source, clove, name)

@spec route_inbound({String.t(), non_neg_integer()}, Clove.t(), atom()) :: :ok

Route an inbound clove to the appropriate destination.

route_outbound(dest, clove, name)

@spec route_outbound({String.t(), non_neg_integer()}, Clove.t(), map()) :: :ok

Route data to the appropriate destination.

route_outbound_discovery(lap2_addr, clove, name)

@spec route_outbound_discovery(String.t(), Clove.t(), atom()) :: :ok

route_tcp(query, name)

@spec route_tcp(Query.t(), atom()) :: :ok

save_query_relay(query_id, proxy_seq, name)

@spec save_query_relay(non_neg_integer(), non_neg_integer(), atom()) :: :ok

Save the origin relay sequence number for a given query ID. This is done so responses can be routed back to the requesting node.

start_link(config)

@spec start_link(map()) :: GenServer.on_start()

Start the Router process.

terminate(reason, arg2)

@spec terminate(any(), map()) :: :ok

Terminate the UDP server.

update_config(config, name)

@spec update_config(map(), atom()) :: :ok

Update the configuration in the state