LAP2.Services.FileIO (LAP2 v0.1.0)

Example service provider for file transfer.

Functions

handle_call(arg1, from, state)

@spec handle_call({:register, String.t(), String.t(), String.t()}, any(), map()) ::
  {:reply, :ok, map()}

@spec handle_call(:get_state, any(), map()) :: {:reply, map(), map()}

@spec handle_call({:set_status, map(), String.t()}, any(), map()) ::
  {:reply, :ok, map()}

@spec handle_call({:get_status, String.t()}, any(), map()) ::
  {:reply, {:ok, map()} | {:error, atom()}, map()}

@spec handle_call(:get_registry_table, any(), map()) :: {:reply, map(), map()}

init(registry_table)

@spec init(map()) :: {:ok, map()}

Initialise the data handler GenServer.

parse_cmd(json_data, serv_id)

@spec parse_cmd(map() | binary(), String.t()) :: :ok | :error | {:error, atom()}

Parse a JSON request/response for a file and return the file contents.

request_file(intro_points, remote_serv_id, filename, save_location, serv_id, fragment_idx \\ 0)

@spec request_file(
  [String.t()],
  String.t(),
  String.t(),
  String.t(),
  String.t(),
  non_neg_integer()
) ::
  :ok | :error

Request a file from a remote FileIO service. Provide a list of introduction points to the remote service.

run_service(master_name)

@spec run_service(atom()) :: {:ok, String.t()} | :error

Register and start the File IO listener service.

start_link(config)

@spec start_link(map()) :: GenServer.on_start()

Start the Master process.

stop_service(serv_id)

@spec stop_service(String.t()) :: :ok | :error

Stop the service.