LAP2.Utils.ProtoBuf.RequestHelper (LAP2 v0.1.0)

Helper functions for processing LAP2 requests. Verifies integrity, serves as interface for serialising/deserialising with ProtoBuf.

Functions

build_fin_crypto(ring_sign)

@spec build_fin_crypto(binary()) :: {:fin_ke, KeyExchangeFinal.t()}

Build the key exchange finalisation struct

build_init_crypto(identity, ephem_pk, generator, ring_pk, signature)

@spec build_init_crypto(binary(), binary(), binary(), binary(), binary()) ::
  {:init_ke, KeyExchangeInit.t()}

Build the key exchange initialisation struct

build_key_rotation(new_key)

@spec build_key_rotation(binary()) :: {:key_rot, KeyRotation.t()}

Build a key rotation request struct

build_request(crypto, data, req_type, req_id)

@spec build_request({atom(), map()}, binary(), binary(), non_neg_integer()) ::
  Request.t()

build_resp_crypto(identity, ephem_pk, generator, ring_pk, signature, ring_sign)

@spec build_resp_crypto(binary(), binary(), binary(), binary(), binary(), binary()) ::
  {:resp_ke, KeyExchangeResponse.t()}

Build the key exchange response struct

build_symmetric_crypto()

@spec build_symmetric_crypto() :: {:sym_key, SymmetricKey.t()}

Build a regular symmetric key struct Note: HMAC key is set when the HMAC is computed

deserialise(data, struct)

@spec deserialise(binary(), Request | EncryptedRequest) ::
  {:ok, Request.t() | EncryptedRequest.t()} | {:error, any()}

Deserialise a request struct.

deserialise_and_unwrap(enc_request)

@spec deserialise_and_unwrap(binary()) :: {:ok, Request.t()} | {:error, atom()}

Deserialise to EncryptedRequest then finally to Request struct. If unencrypted, deserialise wrapped data. If encrypted, unwrap and decrypt request.

deserialise_and_unwrap(enc_request, proxy_seq, crypto_mgr_name \\ :crypto_manager)

@spec deserialise_and_unwrap(binary(), integer(), atom()) ::
  {:ok, Request.t()} | {:error, atom()}

encrypt_and_wrap(request, proxy_seq, crypto_mgr_name \\ :crypto_manager)

@spec encrypt_and_wrap(Request.t(), non_neg_integer(), atom()) ::
  {:ok, EncryptedRequest.t()} | {:error, atom()}

Encrypt a Request struct and serialise to EncryptedRequest struct.

gen_finalise_exchange(request, proxy_seq, clove_seq, crypto_mgr \\ :crypto_manager)

@spec gen_finalise_exchange(Request.t(), non_neg_integer(), non_neg_integer(), atom()) ::
  {:ok, EncryptedRequest.t()} | {:error, atom()}

Finish a key exchange with a remote proxy. Different from recv_finalise_exchange as this is called by the initiator.

gen_response(request, proxy_seq, crypto_mgr \\ :crypto_manager)

@spec gen_response(Request.t(), non_neg_integer(), atom()) ::
  {:ok, EncryptedRequest.t()} | {:error, atom()}

Respond to a key exchange request from a remote proxy.

init_exchange(clove_seq, crypto_mgr \\ :crypto_manager)

@spec init_exchange(non_neg_integer(), atom()) ::
  {:ok, EncryptedRequest.t()} | {:error, atom()}

Initiate a key exchange with a remote proxy.

init_key_rotation(proxy_seq, crypto_mgr \\ :crypto_manager)

@spec init_key_rotation(non_neg_integer(), atom()) ::
  {:ok, EncryptedRequest.t()} | {:error, atom()}

Initiate key rotation with a remote proxy.

recv_finalise_exchange(request, proxy_seq, crypto_mgr \\ :crypto_manager)

@spec recv_finalise_exchange(Request.t(), non_neg_integer(), atom()) ::
  {:ok, EncryptedRequest.t()} | {:error, :invalid_signature}

Finish a key exchange with a remote proxy. Different from gen_finalise_exchange as this is called by the receiver proxy.

rotate_keys(request, proxy_seq, crypto_mgr \\ :crypto_manager)

@spec rotate_keys(Request.t(), non_neg_integer(), atom()) ::
  {:ok, EncryptedRequest.t()} | {:error, :invalid}

Perform key rotation request with a remote proxy.

serialise(request)

@spec serialise(Request.t() | EncryptedRequest.t()) ::
  {:ok, binary()} | {:error, any()}

Serialise a request struct.

wrap_unencrypted(request)

@spec wrap_unencrypted(Request.t()) :: {:ok, EncryptedRequest.t()} | {:error, atom()}